fix(useimageupload): refactor into ordinary functions
